Moved the SemaLock
