+++ Manager/Classes/ClientConnectionHandler.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using Logging;
 using Manager.Classes.Chat;
 using Manager.Classes.Toast;
+using Shared.Classes.Threading;
 using Storage.Classes;
 using Storage.Classes.Contexts;
 using Storage.Classes.Events;
