+++ Manager/Manager.csproj
@@ -125,7 +125,6 @@
     <Compile Include="Classes\Chat\ChatDataTemplate.cs" />
     <Compile Include="Classes\Chat\ObservableChatDictionaryList.cs" />
     <Compile Include="Classes\Chat\SaveObservableChatDictionaryList.cs" />
-    <Compile Include="Classes\Chat\SemaLock.cs" />
     <Compile Include="Classes\Client.cs" />
     <Compile Include="Classes\ClientConnectionHandler.cs" />
     <Compile Include="Classes\ConnectionHandler.cs" />
