+++ UWPX_UI/App.xaml.cs
@@ -10,6 +10,7 @@ using Manager.Classes.Toast;
 using Push.Classes;
 using Push.Classes.Events;
 using Shared.Classes;
+using Shared.Classes.Threading;
 using Storage.Classes;
 using Storage.Classes.Contexts;
 using Storage.Classes.Models.Chat;
